# server/src/docs/openapi.yaml
openapi: 3.0.3
info:
  title: JobBoard API
  description: Full-featured Job Board Platform API
  version: 1.0.0
  contact:
    name: JobBoard Team
servers:
  - url: http://localhost:3001/api
    description: Local development
  - url: http://localhost/api
    description: Docker production
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...
paths:
...
  /users/me:
    get:
      tags: [Users]
      summary: Get own profile
      security:
        - bearerAuth: []
      responses:
        '200':
          description: User profile
    put:
      tags: [Users]
      summary: Update profile
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                firstName:
                  type: string
                lastName:
                  type: string
                avatarUrl:
                  type: string
                resumeUrl:
                  type: string
      responses:
        '200':
          description: Profile updated
...
  /admin/users:
    get:
      tags: [Admin]
      summary: List all users
      security:
        - bearerAuth: []
      responses:
        '200':
          description: List of users
  /admin/users/{id}/role:
    patch:
      tags: [Admin]
      summary: Change user role
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [role]
              properties:
                role:
                  type: string
                  enum: [ADMIN, EMPLOYER, SEEKER]
      responses:
        '200':
          description: Role updated
...